Move half of sshcommon.c out into sshutils.c.
Like other 'utils' modules, the point is that sshutils.c has no
external dependencies, so it's safe to include in a tool without
requiring you to bring in a cascade of other modules you didn't really
want.

Right now I'm only planning to use this change in an out-of-tree
experiment, but it's harmless to commit the change itself here.